feat: add edit skill hooks
